Cleanup, Improve, Fix typos, add plugin specific msg
Add PluginMsg std::string for all outputs.
Add enum for itemsizes..don't know if that was already a thing elsewhere, but now it's here for me.
HandlePersonalLoot: Changed to not be used if AutoLootAllIsOn is on. Had it backwards.
Add Enum for use with DoSharedAdvLootAction so it's clear what I'm setting it to do.
Fix bracketing in assigning PLOOTITEM which actually makes it work correctly.
Removed unused IWant hold-over from when I was looking at plure logic. 
Remove unnecessary comment about making an enum.
Add what we're doing as output spam to MQ2 Window. Probably should make this optional.
Put the pause check in OnPulse, so now it actually does something.
Now checking to ensure we have space to loot an item. Default is just see if we have an empty giant slot. Will need to be improved.
Relocated pChar and pChar2 assignments to prior to pAdvLoot
Remove debug statements that I've literally never triggered.
Reduced indentation in OnPulse for readability.