Allow use of PersonalLootList, Add Enum for personal loot...
Being Master looter doesn't stop you from managing personal loot list with plugin. 
If you are the Master Looter don't do anything in the personal list. If you aren't, and there is a filter, let EQ manage it. If you aren't and there isn't a filter, just grab the item so it doesn't rot? Kinda counter to method we're doing if you're master looter, so not sure it will stay this way.
Remove commented out alternate method of getting shared loot items. Couldn't iterate through the list that way anyhow.