Fix output. Fix no group. Fix not counting group members out of zone
When you ran out of bag space, the output message was formatted, but failed to provide the string for the plugin msg. 

Counting Group members was only counting group members if they were in the zone with you. Which would cause it to fail to loot if you were in a zone by yourself. 

When you were by yourself, it wouldn't autoloot. Fixt.